+++ methods/response_upgradeinventoryitem.js
@@ -1,9 +1,9 @@
-exports.run = (method, sign, path, fs, tools, id, request, config) => {
+exports.run = (method, sign, path, fs, tools, id, db, config) => {
 	var file = fs.readFileSync(path, 'utf8')
 	var new_file = file.substr(0, file.indexOf('!'))
 	fs.writeFileSync(path, new_file, 'utf8')
 	var response_JSON = require("D:/TotalDominationHelper/"+path)//в будущем упростить
-	var options = {
+	/*var options = {
   uri: config.url+"/upgr",
   method: 'POST',
   json: {
@@ -21,7 +21,17 @@ request(options, function (error, response, body) {
   }else {
     console.log(error)
   }
-});
+});*/
+ if(!db.get('todos').find({sign: sign}).value()){
+    var num = db.get('last_num').value()
+    db.get('todos')
+      .push({num: num, type: method, id: 0, time_start: 0, sign: sign, id_p: id, time_end: response_JSON.o.t}).write()
+      db.update('last_num', n => n + 1).write()
+  }else {
+    db.get('todos')
+      .find({sign: sign})
+      .assign({time_end: response_JSON.o.t}).write()
+  }
 	tools.log("Улучшение предмета завершится в:", tools.convertTimestamp(response_JSON.o.t))
 	fs.removeSync(path, { recursive: true });
 };
